Reduce the memory complexity
Maintain a scalar sum, rather a vector of sets of chars.  This reduces
memory consumption from O(N * M_avg), where N is the number of groups
and M_avg is the average number of answers per group, to O(M_max), where
M_max is the maximum number of answers in any group.