Use dynamic Circle type for day 23 part 1
It still seems to give the wrong answer for part 2.  The static version
(using compile-time constants and raw arrays) immediately blew the stack
when I tried it on part 2, was slow (onced I switched to a vector)
because it relied heavily on full rotations of the array, and was very
difficult to debug (because it's hard to change the problem's parameters
on the fly).