Solve Day 14 Part 2
It's off by one for my input, but not for the samples.  I clearly have
an off-by-one error, so I submitted a guess that was one less than the
actual return value of my solver function, and that was correct.  I'll
debug this properly at some point, but at the moment, it does actually
print the correct answers for all samples and both my inputs.